feat (frontend): remove fab's unecessary re-renders
