--- frontend/src/views/components/app/Fab/index.tsx
@@ -1,54 +1,0 @@
-import { Plus } from 'lucide-react';
-
-import { cn } from '@app/lib/utils';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from '@views/components/ui/DropdownMenu';
-
-import { useFabController } from './useFabController';
-
-export function Fab() {
-  const { fabItems } = useFabController();
-
-  return (
-    <DropdownMenu modal={false}>
-      <DropdownMenuTrigger asChild>
-        <button
-          type="button"
-          className="fixed bottom-4 right-4 md:bottom-6 md:right-6 text-white bg-teal-900 dark:bg-primary transition-colors rounded-full p-3 outline-none"
-        >
-          <Plus className="size-6" />
-        </button>
-      </DropdownMenuTrigger>
-
-      <DropdownMenuContent
-        align="end"
-        className="!shadow-default space-y-2 px-2 py-2.5 bg-white dark:bg-popover border-transparent dark:border-accent"
-      >
-        {fabItems.map(({ label, icon: Icon, bgIcon, colorIcon, handler }) => (
-          <DropdownMenuItem
-            key={label}
-            className="cursor-pointer"
-            onSelect={handler}
-          >
-            <div
-              className={cn(
-                'flex items-center justify-center p-1.5 rounded-full',
-                bgIcon,
-              )}
-            >
-              <Icon className={cn('size-5', colorIcon)} />
-            </div>
-
-            <span className="text-sm text-gray-800 dark:text-foreground tracking-[-0.5px]">
-              {label}
-            </span>
-          </DropdownMenuItem>
-        ))}
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
