+++ frontend/src/views/layouts/AppLayout/components/GlobalModals/index.tsx
@@ -1,5 +1,4 @@
 import { NewCustomerModal } from '@views/modals/NewCustomerModal';
-import { NewOrderModal } from '@views/modals/NewOrderModal';
 import { NewPaymentModal } from '@views/modals/NewPaymentModal';
 
 import { useGlobalModalsController } from './useGlobalModalsController';
@@ -7,10 +6,8 @@ import { useGlobalModalsController } from './useGlobalModalsController';
 export function GlobalModals() {
   const {
     isOpenNewCustomerModal,
-    isOpenNewOrderModal,
     isOpenNewPaymentModal,
     handleCloseNewCustomerModal,
-    handleCloseNewOrderModal,
     handleCloseNewPaymentModal,
   } = useGlobalModalsController();
 
@@ -18,9 +15,6 @@ export function GlobalModals() {
     <>
       {isOpenNewCustomerModal && (
         <NewCustomerModal open onClose={handleCloseNewCustomerModal} />
-      )}
-      {isOpenNewOrderModal && (
-        <NewOrderModal open onClose={handleCloseNewOrderModal} />
       )}
       {isOpenNewPaymentModal && (
         <NewPaymentModal open onClose={handleCloseNewPaymentModal} />
